+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Updated
 
-- `gradle -> 8.10.1`
+- `gradle -> 8.10.2`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.11`
 
 ## [0.7.0] - 2024-08-16