+++ CHANGELOG.md
@@ -10,6 +10,14 @@
 
 ### Fixed
 
+### Removed
+
+### Updated
+
+## [0.7.1] - 2024-09-26
+
+### Fixed
+
 - commit number not being null
 
 ### Removed
@@ -119,7 +127,9 @@
 - `org.eclipse.jgit:org.eclipse.jgit -> 6.5.0.202303070854-r`
 - `com.javiersc.semver:semver-core -> 0.1.0-beta.13`
 
-[Unreleased]: https://github.com/JavierSegoviaCordoba/semver-gradle-plugin/compare/0.7.0...HEAD
+[Unreleased]: https://github.com/JavierSegoviaCordoba/semver-gradle-plugin/compare/0.7.1...HEAD
+
+[0.7.1]: https://github.com/JavierSegoviaCordoba/semver-gradle-plugin/compare/0.7.0...0.7.1
 
 [0.7.0]: https://github.com/JavierSegoviaCordoba/semver-gradle-plugin/compare/0.6.2...0.7.0